requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# API/crawler.py: 4
# API/src/fetch_facts.py: 3
PyYAML == 6.0

# API/crawler.py: 7
# API/src/fact_checker.py: 4
# API/src/fetch_facts.py: 6
pandas == 1.3.5

# API/src/translator.py: 2
protobuf == 3.20.1

# API/src/fact_checker.py: 7
python_dotenv == 0.20.0

# API/src/fact_checker.py: 3
requests == 2.28.1

# API/src/translator.py: 3
six == 1.16.0

# API/src/fetch_facts.py: 4
tqdm == 4.62.2
